Move dataview inside viz folder
