+++ src/lib/viz/dataview/category/CategoryDataView.spec.ts
@@ -1,6 +1,6 @@
-import { Layer } from '../../viz/layer/Layer';
+import { Layer } from '../../layer/Layer';
 import { CategoryDataView } from './CategoryDataView';
-import { AggregationType } from '../../data/operations/aggregation/aggregation';
+import { AggregationType } from '../../../data/operations/aggregation/aggregation';
 import { CartoDataViewError, dataViewErrorTypes } from '../DataViewError';
 import { DataViewCalculation } from '../mode/DataViewMode';
