+++ src/lib/viz/dataview/category/CategoryDataView.ts
